requirements.txt
equinox==0.13.6
optax==0.2.7
chex==0.1.91
pillow
jaxtyping
pytest
matplotlib
wandb
xprof